requirements.txt
pylottie==0.2.0
Pillow==10.1.0
moviepy==1.0.3